Refaktorera för bättre överblick och förståelse
