+++ v02/bin/v02.dart
@@ -2,54 +2,31 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:v02/v02.dart' as v02;
+import 'package:v02/v02_helpers.dart' as v02_helpers;
+import 'package:v02/v02_menuOptions.dart' as v02_menu;
 
 void main(List<String> arguments) {
   gameHandler();
 }
 
-String mainMenu() {
-  print("1. Skapa hjälte");
-  print("2. Visa alla hjältar");
-  print("3. Avsluta");
-  stdout.write("Välj alternativ: ");
-  String input = stdin.readLineSync() ?? "";
-  return input;
-}
-
-String listHeroesMenu() {
-  clearTerminal();
-  print("1. Tillbaka");
-  stdout.write("Välj alternativ: ");
-  String input = stdin.readLineSync() ?? "";
-  if (input.isNotEmpty) clearTerminal();
-  return input;
-}
 
 
-void clearTerminal() {
-  if (Platform.isWindows) {
-    stdout.write('\x1B[2J\x1B[0;0H');
-  } else {
-    print(Process.runSync("clear", [], runInShell: true).stdout);
-  }
-}
 
 
 void gameHandler() {
   bool isRunning = true;
   String input = "";
-  clearTerminal();
+  v02_helpers.clearTerminal();
 
   while(isRunning) {
     switch(input) {
       case "2":
-        input = listHeroesMenu();
+        input = v02_menu.listHeroesMenu();
       case "3":
         exit(0);
       default:
-        input = mainMenu();
+        input = v02_menu.mainMenu();
         break;
-      
     }
   }
 }
