+++ v02/lib/v02_menuOptions.dart
@@ -1,0 +1,21 @@
+
+import 'dart:io';
+import 'package:v02/v02_helpers.dart' as v02_helpers;
+
+String mainMenu() {
+  print("1. Skapa hjälte");
+  print("2. Visa alla hjältar");
+  print("3. Avsluta");
+  stdout.write("Välj alternativ: ");
+  String input = stdin.readLineSync() ?? "";
+  return input;
+}
+
+String listHeroesMenu() {
+  v02_helpers.clearTerminal();
+  print("1. Tillbaka");
+  stdout.write("Välj alternativ: ");
+  String input = stdin.readLineSync() ?? "";
+  if (input.isNotEmpty) v02_helpers.clearTerminal();
+  return input;
+}
